requirements.txt
fastapi[standard]
pydantic-settings
sqlmodel
alembic
psycopg2-binary